remove write/deleteDiocesanCalendar.php
this functionality is now well defined in the API, and should not be handled on the frontend. The frontend will call the API to add or remove a diocesan calendar. Also move any included php scripts into the includes folder, so that pages will stand out clearly compared to included files