focus first tab with data
when loading a diocesan calendar in the `extending` section, focus the carousel on the first card that has a non empty `name` input. When the first card has no inputs with any vals, it almost seems like no data was loaded. By focusing on the first carousel card that has data, it is immediately obvious that data has been loaded