Begin migration to TailwindCSS
The [sb-admin](https://startbootstrap.com/template/sb-admin) theme was a
great way to quickly get this project off the ground, but as the project
matures we're discovering several downsides:

 - The CDN approach makes a very simple dev experience, but doesn't
   allow us to easily customize the colors.

 - sb-admin provides a complex interface for dashboards that need to
   display a lot of information, but we don't have that much info to
   display. (E.g. we don't need a navbar and a side panel with
   partially-duplicated links.) A simpler interface is less overwhelming
   for new users.

 - Tailwind makes it easy to build a simple theme ourselves and
   customize it as needed.

This commit introduces TailwindCSS to replace Bootstrap and sb-admin,
with the goal of retaining as much simplicity (lightweight tooling and a
simple build process) as possible.

The migration is not complete, but we have a navbar and part of
index.html re-skinned -- enough to start to see what this might look
like.