allow loading and editing i18n data for calendars # Issue being worked on #161
# Summary of changes
With the new backend API functionality being worked on in Liturgical-Calendar/LiturgicalCalendarAPI/issues/284 / Liturgical-Calendar/LiturgicalCalendarAPI/pull/285, we load i18n data for the current calendar so as to be able to PUT / PATCH i18n data from the frontend interfaces.

While working on these implementations, I also took the opportunity to refactor a lot of the code from jQuery to native javascript, considering that native javascript can do pretty much everything that jQuery has offered up until now. Even bootstrap no longer requires jQuery. The only bootstrap plugin that is still a jQuery plugin is the `multiselect`, so we still need to use jQuery for this.